gunzip hg19 repeatmasker output and generate pro-A/pro-B repseq bed files

+++ init.sh
@@ -10,7 +10,10 @@ if [[ ! -e hg19/README ]]; then
 	w3m -dump http://hgdownload.soe.ucsc.edu/goldenPath/hg19/bigZips/ >hg19/README
 fi
 # hg19 repeatmasker output
-wget -nc -O hg19/hg19.fa.out.gz 'ftp://hgdownload.soe.ucsc.edu/goldenPath/hg19/bigZips/hg19.fa.out.gz'
+if [[ ! -f hg19/hg19.fa.out ]]; then
+	wget -nc -O hg19/hg19.fa.out.gz 'ftp://hgdownload.soe.ucsc.edu/goldenPath/hg19/bigZips/hg19.fa.out.gz'
+	gunzip hg19/hg19.fa.out.gz
+fi
 # hg19 chromosome info (for windowing)
 if [[ ! -f hg19/hg19.txt ]]; then
 	mysql --user=genome --host=genome-mysql.cse.ucsc.edu -A -e "select chrom, size from hg19.chromInfo" >hg19/hg19.txt
@@ -27,17 +30,26 @@ fi
 # download files specified in csv's, convert excel stuff
 Rscript init.R
 
-# extract repseq subclasses
+# extract pro-A repseqs
 awk '
-NR > 1{
+FNR == NR && NR > 1{
 	if($4 > 0.04)
-		print $0
+		s[$3] = ""
 }
-' gf/huvec.repseq.tsv > gf/huvec.proa.tsv
+FNR != NR{
+	if($10 in s)
+		printf("%s\t%s\t%s\t%s\n", $5, $6, $7, $10)
+}
+' gf/huvec.repseq.tsv hg19/hg19.fa.out > cnt/huvec.repseq.proa.bed
 
+# extract pro-B repseqs
 awk '
-NR > 1{
+FNR == NR && NR > 1{
 	if($4 < -0.01)
-		print $0
+		s[$3] = ""
 }
-' gf/huvec.repseq.tsv > gf/huvec.prob.tsv
+FNR != NR{
+	if($10 in s)
+		printf("%s\t%s\t%s\t%s\n", $5, $6, $7, $10)
+}
+' gf/huvec.repseq.tsv hg19/hg19.fa.out > cnt/huvec.repseq.prob.bed
